# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  topicsLearned: number;
  practiceSessions: number;
  savedItems: number;
  planProgress: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  market: 'market',
  study: 'study',
  chat: 'chat',
  safehelp: 'safehelp',
  plan: 'plan',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  detail: string;
  time: string;
}

export type MarketAnalysisInputCurrency = typeof MarketAnalysisInputCurrency[keyof typeof MarketAnalysisInputCurrency];


export const MarketAnalysisInputCurrency = {
  UAH: 'UAH',
  USD: 'USD',
  EUR: 'EUR',
  GBP: 'GBP',
  PLN: 'PLN',
} as const;

export interface MarketAnalysisInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 0 */
  price: number;
  currency: MarketAnalysisInputCurrency;
  /** @nullable */
  imageName?: string | null;
}

export type MarketAnalysisVerdict = typeof MarketAnalysisVerdict[keyof typeof MarketAnalysisVerdict];


export const MarketAnalysisVerdict = {
  good: 'good',
  fair: 'fair',
  expensive: 'expensive',
} as const;

export interface MarketAnalysis {
  category: string;
  brand: string;
  model: string;
  description: string;
  condition: string;
  estimatedPriceMin: number;
  estimatedPriceMax: number;
  currency: string;
  sellerPrice: number;
  difference: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  thingsToCheck: string[];
  verdict: MarketAnalysisVerdict;
  demo: boolean;
}

export type StudyExplanationInputDepth = typeof StudyExplanationInputDepth[keyof typeof StudyExplanationInputDepth];


export const StudyExplanationInputDepth = {
  simple: 'simple',
  standard: 'standard',
  deep: 'deep',
} as const;

export interface StudyExplanationInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  level: string;
  /** @minLength 1 */
  topic: string;
  depth: StudyExplanationInputDepth;
}

export interface StudyExplanation {
  topic: string;
  whatIsIt: string;
  keyRule: string;
  example: string;
  memoryTip: string;
  commonMistakes: string[];
  checkQuestion: string;
  demo: boolean;
}

export type ChatRequestMode = typeof ChatRequestMode[keyof typeof ChatRequestMode];


export const ChatRequestMode = {
  study: 'study',
  market: 'market',
  general: 'general',
  safehelp: 'safehelp',
} as const;

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  /** @minLength 1 */
  message: string;
  mode: ChatRequestMode;
  history?: ChatMessage[];
}

export interface ChatResponse {
  message: string;
  demo: boolean;
}

export interface SafeHelpRequest {
  /** @minLength 1 */
  situation: string;
  /** @nullable */
  category?: string | null;
}

export type SafeHelpResponseRiskLevel = typeof SafeHelpResponseRiskLevel[keyof typeof SafeHelpResponseRiskLevel];


export const SafeHelpResponseRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  emergency: 'emergency',
} as const;

export interface SafeHelpResponse {
  situation: string;
  riskLevel: SafeHelpResponseRiskLevel;
  summary: string;
  immediateActions: string[];
  avoid: string[];
  whenToGetHelp: string;
  recommendedContact: string;
  emergency: boolean;
  demo: boolean;
}