# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { ChatRequestMode } from './chatRequestMode';

export interface ChatRequest {
  /** @minLength 1 */
  message: string;
  mode: ChatRequestMode;
  history?: ChatMessage[];
}