# lib/api-zod/src/generated/types/marketAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAnalysisVerdict } from './marketAnalysisVerdict';

export interface MarketAnalysis {
  category: string;
  brand: string;
  model: string;
  description: string;
  condition: string;
  estimatedPriceMin: number;
  estimatedPriceMax: number;
  currency: string;
  sellerPrice: number;
  difference: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  thingsToCheck: string[];
  verdict: MarketAnalysisVerdict;
  demo: boolean;
}