# lib/api-zod/src/generated/types/marketAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAnalysisInputCurrency } from './marketAnalysisInputCurrency';

export interface MarketAnalysisInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 0 */
  price: number;
  currency: MarketAnalysisInputCurrency;
  /** @nullable */
  imageName?: string | null;
}