# lib/api-zod/src/generated/types/safeHelpRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SafeHelpRequest {
  /** @minLength 1 */
  situation: string;
  /** @nullable */
  category?: string | null;
}