# lib/api-zod/src/generated/types/safeHelpResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SafeHelpResponseRiskLevel } from './safeHelpResponseRiskLevel';

export interface SafeHelpResponse {
  situation: string;
  riskLevel: SafeHelpResponseRiskLevel;
  summary: string;
  immediateActions: string[];
  avoid: string[];
  whenToGetHelp: string;
  recommendedContact: string;
  emergency: boolean;
  demo: boolean;
}