# lib/api-zod/src/generated/types/studyExplanation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyExplanation {
  topic: string;
  whatIsIt: string;
  keyRule: string;
  example: string;
  memoryTip: string;
  commonMistakes: string[];
  checkQuestion: string;
  demo: boolean;
}