# lib/api-zod/src/generated/types/studyExplanationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyExplanationInputDepth } from './studyExplanationInputDepth';

export interface StudyExplanationInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  level: string;
  /** @minLength 1 */
  topic: string;
  depth: StudyExplanationInputDepth;
}